requirements.txt
PyPDF2==3.0.1
pdf2image==1.16.3
pytesseract==0.3.10
Pillow==10.0.0

tensorflow==2.13.0
keras==2.13.1

tensorflow-hub==0.13.0

numpy
scikit-learn